agentic_ai/requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn[standard]
joblib
requests
faiss-cpu
sentence-transformers
catboost
apscheduler
pydantic